% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyth4.R
\name{pyth4}
\alias{pyth4}
\title{hw06 pythagoream theorem}
\usage{
pyth4(sa, sb, sc)
}
\arguments{
\item{sa}{a number}

\item{sb}{a number}

\item{sc}{a number}
}
\value{
the third line of a triangle
}
\description{
hw06 pythagoream theorem
}
\keyword{pythagoream}
\keyword{theorem}
