% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trmean.R
\name{trmean}
\alias{trmean}
\title{hw06 trim mean}
\usage{
trmean(x, s, l)
}
\arguments{
\item{x}{a list of number}

\item{s}{the small number}

\item{l}{the large number}
}
\value{
trimmed mean
}
\description{
hw06 trim mean
}
\keyword{mean}
\keyword{trimmed}
